Factor out name mempool to separate files.
The main Namecoin code is split into two files, names/common and
names/main.  These files are monolithic, which is not very nice for
maintenance and development.

This change splits out the mempool-related code (the CNameMemPool class
and the conflict tracker) into its own set of source files; this will be
useful in a follow-up change for
https://github.com/namecoin/namecoin-core/issues/304.

All changes here are just moves of code (and updates to the #includes
in all new locations as needed).
