+++ src/Makefile.am
@@ -154,6 +154,7 @@ BITCOIN_CORE_H = \
   names/common.h \
   names/encoding.h \
   names/main.h \
+  names/mempool.h \
   net.h \
   net_processing.h \
   netaddress.h \
@@ -278,6 +279,7 @@ libbitcoin_server_a_SOURCES = \
   dbwrapper.cpp \
   miner.cpp \
   names/main.cpp \
+  names/mempool.cpp \
   net.cpp \
   net_processing.cpp \
   node/coin.cpp \
